Patch on fixed in delete function and 'other email' sending of lead

+++ src/Controller/UserLeadsController.php
@@ -225,8 +225,8 @@ class UserLeadsController extends AppController
 
                       foreach($other_email as $oekey => $em) {
 
-                        if( !empty($em) || $em != '') {
-                            $other_email_to_add = ltrim($em);
+                        if (trim($em) != '') {
+                            $other_email_to_add = $em; //ltrim($em);
                             $users_email[$other_email_to_add] = $other_email_to_add;  
                         }
 
@@ -345,8 +345,8 @@ class UserLeadsController extends AppController
 
                       foreach($other_email as $oekey => $em) {
 
-                        if( !empty($em) || $em != '') {
-                            $other_email_to_add = ltrim($em);
+                        if (trim($em) != '') {
+                            $other_email_to_add = $em; //ltrim($em);
                             $users_email[$other_email_to_add] = $other_email_to_add;  
                         }
 
@@ -414,23 +414,33 @@ class UserLeadsController extends AppController
         $p = $this->default_group_actions;
         if( $p && $p['leads'] != 'View, Edit and Delete' ){
             return $this->redirect(['controller' => 'users', 'action' => 'no_access']);
-        } 
-              
-        $login_user_id = $this->user->id;
-        $lead_lock = $this->Leads->get($id, [ 'contain' => ['LastModifiedBy'] ]);         
-
-        if($lead_lock->is_lock && $lead_lock->last_modified_by->id != $this->user->id) {
-          $this->Flash->error(__('This lead is being accessed by another user, please try again later.'));
-          return $this->redirect(['action' => 'index']);
-        }       
-
-        $this->request->allowMethod(['post', 'delete']);
-        $lead = $this->Leads->get($id);
-        if ($this->Leads->delete($lead)) {
-            $this->Flash->success(__('The lead has been deleted.'));
-        } else {
-            $this->Flash->error(__('The lead could not be deleted. Please, try again.'));
         }
+
+        $lead_exists = $this->Leads->exists(['id' => $id]);
+        if($lead_exists) {
+
+          $login_user_id = $this->user->id;
+          $lead_lock = $this->Leads->get($id, [ 'contain' => ['LastModifiedBy'] ]);         
+
+          if($lead_lock->is_lock && $lead_lock->last_modified_by->id != $this->user->id) {
+            $this->Flash->error(__('This lead is being accessed by another user, please try again later.'));
+            return $this->redirect(['action' => 'index']);
+          }       
+
+          $this->request->allowMethod(['post', 'delete']);
+          $lead = $this->Leads->get($id);
+          if ($this->Leads->delete($lead)) {
+              $this->Flash->success(__('The lead has been deleted.'));
+          } else {
+              $this->Flash->error(__('The lead could not be deleted. Please, try again.'));
+          }
+
+        } else {
+
+          $this->Flash->error(__('The lead could not be found. Please try again.'));
+
+        }
+
         return $this->redirect(['action' => 'index']);
     }
